[Web UI] Decrease counter size.

+++ src/bosixnet-webui/bosixnet-webui.cpp
@@ -102,7 +102,7 @@ int main(int argc, char **argv)
     read_hosts();
     read_timestamps();
 
-    unsigned long long counter = 0;
+    unsigned long counter = 0;
     while (FCGI_Accept() >= 0) {
         ++counter;
